Fix base for device offset in PairPos format 1.

+++ src/ot/layout/dyn_array.rs
@@ -15,7 +15,7 @@ impl<'a> DynArray<'a> {
         count: usize,
         stride: usize,
     ) -> Option<Self> {
-        let len = usize::from(count) * stride;
+        let len = count * stride;
         s.read_bytes(len).map(|data| Self { data, stride })
     }
